requirements.txt
# Harvester II Trading System Dependencies

aiohttp>=3.9.0
alpha-vantage>=2.3.1
bandit>=1.7.0

# Caching
cachetools>=5.3.0
commitizen>=3.29.0
hmmlearn>=0.3.0

# Logging and monitoring
loguru>=0.7.0
mypy>=1.11.0
numpy>=1.24.0

# Hyperparameter optimization
optuna>=3.0.0
pandas>=2.0.0
polars>=0.20.0,<1.0.0
pre-commit>=3.0.0
prometheus-client>=0.20.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Configuration and environment
python-dotenv>=1.0.0

# Google Trends data
pytrends>=4.9.2
# sqlcipher3>=0.5.0  # Requires C++ build tools, install separately if needed
# sqlite3 is built-in with Python - no installation needed

# API clients
requests>=2.31.0
responses>=0.24.0

# Development and code quality
ruff>=0.6.0

# Task scheduling
schedule>=1.2.0

# Machine learning and probabilistic modeling
scikit-learn>=1.3.0
scipy>=1.10.0

# Notifications
# smtplib is built-in with Python - no installation needed
slack-sdk>=3.21.0

# Database
SQLAlchemy>=2.0.0

# Technical indicators
ta>=0.10.2

# Advanced backtesting
vectorbt>=0.28.0
# Core data and analysis
yfinance>=0.2.18
